[spirv-writer] Add array type generation.
This CL adds generation of OpTypeArray and OpTypeRuntimeArray to the
SPIR-V writer.

Bug: tint:5
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/17821

+++ src/ast/type/array_type.cc
@@ -20,7 +20,7 @@ namespace type {
 
 ArrayType::ArrayType(Type* subtype) : subtype_(subtype) {}
 
-ArrayType::ArrayType(Type* subtype, size_t size)
+ArrayType::ArrayType(Type* subtype, uint32_t size)
     : subtype_(subtype), size_(size) {}
 
 ArrayType::~ArrayType() = default;
